// src/services/generated/apiTypes/groupResource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */
import type { ApplicantResource } from './applicantResource';

export interface GroupResource {
  id: number;
  name: string;
  /** @nullable */
  message: string | null;
  capacity: number;
  /** @nullable */
  date_time: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  location_link: string | null;
  is_active: boolean;
  /** @nullable */
  last_reminded_at: string | null;
  /** @nullable */
  attendance_closed_at: string | null;
  current_members_count: number;
  applicants?: ApplicantResource[];
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
}